add http response status and fix wrong sql statement

+++ src/main/java/org/aibles/failwall/user/models/Role.java
@@ -1,22 +1,22 @@
 package org.aibles.failwall.user.models;
 
 public class Role {
-    private int id;
+    private Long id;
     private String name;
 
     public Role() {
     }
 
-    public Role(int id, String name) {
+    public Role(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
