+++ src/main/java/org/aibles/failwall/user/repositories/IRoleRepository.java
@@ -3,5 +3,5 @@ package org.aibles.failwall.user.repositories;
 import org.aibles.failwall.user.models.Role;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IRoleRepository extends CrudRepository<Role, Integer> {
+public interface IRoleRepository extends CrudRepository<Role, Long> {
 }
